+++ mleap-xgboost-runtime/src/test/scala/ml/combust/mleap/xgboost/runtime/testing/FloatingPointApproximations.scala
@@ -21,6 +21,4 @@ trait FloatingPointApproximations {
     }
     true
   }
-
 }
-
